Demo2/medical_vision/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
numpy>=1.21.0
pandas>=2.0.0

# Image processing
pydicom>=2.3.0
opencv-python>=4.7.0
albumentations>=1.3.0
scikit-image>=0.20.0
pillow>=9.5.0

# Machine learning utilities
scikit-learn>=1.0.0
scipy>=1.9.0

# Experiment tracking and visualization
wandb>=0.15.0
matplotlib>=3.5.0
seaborn>=0.12.0
tensorboard>=2.12.0

# Data handling and processing
pyarrow>=11.0.0
h5py>=3.8.0
tqdm>=4.65.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=1.0.0

# Development tools
black>=23.3.0  # Code formatting
isort>=5.12.0  # Import sorting
flake8>=6.0.0  # Code linting
pytest>=7.3.0  # Testing
pytest-cov>=4.0.0  # Test coverage

# Optional: CUDA toolkit requirements
# Note: Uncomment if needed for specific CUDA version
# cupy-cuda11x>=11.0.0  # Replace x with your CUDA version
# nvidia-ml-py>=11.525.112

# Performance optimization
numba>=0.56.0